Addressed glorious nits
